+++ main.py
@@ -1,0 +1,9 @@
+from bot.core import bot
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+TOKEN = os.getenv("DISCORD_TOKEN")
+
+if __name__ == "__main__":
+    bot.run(TOKEN)
